requirements.txt
# core dependencies
torch
pandas
datasets
transformers

# progress bars
tqdm

# code formatting
black
isort

# experiment tracking and pipeline versioning
wandb
mlflow
dvc

# user interface
streamlit

# utilities
python-box
pyYAML
ensure
joblib
python-dotenv

-e .